[#17] Refactor naming
