+++ app/src/androidTest/java/com/kks/nimblesurveyjetpackcompose/di/FakeRepoModule.kt
@@ -2,8 +2,8 @@ package com.kks.nimblesurveyjetpackcompose.di
 
 import com.kks.nimblesurveyjetpackcompose.repo.home.HomeRepo
 import com.kks.nimblesurveyjetpackcompose.repo.login.LoginRepo
-import com.kks.nimblesurveyjetpackcompose.ui.presentation.home.HomeScreenKtTest
-import com.kks.nimblesurveyjetpackcompose.ui.presentation.splash.SplashScreenTest
+import com.kks.nimblesurveyjetpackcompose.ui.presentation.home.FakeHomeRepo
+import com.kks.nimblesurveyjetpackcompose.ui.presentation.splash.FakeLoginRepo
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.components.SingletonComponent
@@ -18,9 +18,9 @@ import javax.inject.Singleton
 class FakeRepoModule {
     @Singleton
     @Provides
-    fun provideFakeLoginRepo(): LoginRepo = SplashScreenTest.FakeLoginRepo()
+    fun provideFakeLoginRepo(): LoginRepo = FakeLoginRepo()
 
     @Singleton
     @Provides
-    fun provideFakeHomeRepo(): HomeRepo = HomeScreenKtTest.FakeHomeRepo()
+    fun provideFakeHomeRepo(): HomeRepo = FakeHomeRepo()
 }
