+++ app/src/test/java/com/kks/nimblesurveyjetpackcompose/viewmodel/splash/SplashViewModelTest.kt
@@ -66,7 +66,7 @@ class SplashViewModelTest : BaseViewModelTest() {
         viewModel.login("example@gmail.com", "invalid")
         advanceUntilIdle()
 
-        assertEquals(errorMessage, viewModel.getError.value?.errorMessage)
+        assertEquals(errorMessage, viewModel.error.value?.errorMessage)
     }
 
     @Test
